# openapi.yaml
openapi: 3.0.3
info:
  title: Procadi
  version: 1.0.0
...
paths:
...
  /feedback:
    post:
      tags:
        - feedback
      summary: Add feedback
      operationId: addFeedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewFeedback'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewFeedback'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewFeedback'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Feedback'
...
  /team/findByCompany:
    get:
      tags:
        - team
      summary: Returns teams by company
      operationId: findTeamsByCompany
      parameters:
        - name: company_id
          in: query
          description: The user name for login
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
...
components:
  schemas:
    Address:
      type: object
      properties:
        street:
          type: string
          example: 7 Sheptytskoho
        city:
          type: string
          example: Novoyavorivsk
        country:
          type: string
          example: Ukraine
        zip:
          type: string
          example: '81053'
      xml:
        name: address
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        owner_id:
          $ref: '#/components/schemas/Employee'  
        name:
          type: string
          example: The Boring Team 
      xml:
        name: company
...
    Feedback:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        dateOfCreation:
          type: string
          format: date
          example: '2017-07-21'
        note:
          type: string
        answers:
          type: array
          items: 
            type: string
            example: how are you?
      xml:
        name: feedback
    NewFeedback:
      type: object
      properties:
        dateOfCreation:
          type: string
          format: date
          example: '2017-07-21'
        note:
          type: string
        answers:
          type: array
          items: 
            type: string
            example: how are you?
      xml:
        name: feedback
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        tag:
          type: string
          example: Web
        name:
          type: string
          example: The Boring Team
        participants_id:
          type: array
          items: 
            $ref: '#/components/schemas/EmployeeNoPassword'    
        company_id:
          type: array
          items: 
            $ref: '#/components/schemas/EmployeeNoPassword'  
      xml:
        name: team
...
    Employee:
      type: object
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        addres:
          $ref:  '#/components/schemas/Address'
        email:
          type: string
          example: john@email.com
        company_id:
          $ref:  '#/components/schemas/Company'
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        dateOfBirth:
          type: string
          format: date
          example: '2017-07-21'
        role:
          type: string
          enum:
            - user
            - manager
      xml:
        name: employee
    EmployeeNoPassword:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        addres:
          $ref:  '#/components/schemas/Address'
        email:
          type: string
          example: john@email.com
        company_id:
          $ref:  '#/components/schemas/Company'
        phone:
          type: string
          example: '12345'
        dateOfBirth:
          type: string
          format: date
          example: '2017-07-21'
        role:
          type: string
          enum:
            - user
            - manager
      xml:
        name: employeeNoPassword
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header